// safeTech-mobile/frontend/types/types/type.d.ts
declare interface MarkerData {
    id: number | string; // or string, depending on your ID type
    latitude: number;
    longitude: number;

}
// ...
interface LocationStore {
    userKey: number | null;
    userLatitude: number | null;
    userLongitude: number | null;
    userAddress: string | null;
    destinationLatitude: number | null;
    destinationLongitude: number | null;
    destinationAddress: string | null;
    emergencyId:number | null;
    enroute:string | null;
    destinationLocation: {
        userId: number | null;
        latitude: number | null;
        longitude: number | null;
        address: string | null;
        eId:number | null;
        route:string | null;
    } | null;
    hospitalLocation: {
        userId: number | null;
        latitude: number | null;
        longitude: number | null;
        address: string | null;
        eId:number | null;
        route:string | null;
    } | null;
    setUserLocation: ({
        latitude,
        longitude,
        address,
    }: {
        latitude: number;
        longitude: number;
        address: string;
    }) => void;
    setDestinationLocation: ({
        userId,
        latitude,
        longitude,
        address,
        eId,
        route,
    }: {
        userId: number;
        latitude: number;
        longitude: number;
        address: string;
        eId,
        route:string
    }) => void;
    setHospitalLocation: ({
        userId,
        latitude,
        longitude,
        address,
        eId,
        route,
    }: {
        userId: number;
        latitude: number;
        longitude: number;
        address: string;
        eId,
        route:string
    }) => void;
}

declare interface DriverStore {
    drivers: MarkerData[];
    selectedDriver: number | null;
    setSelectedDriver: (driverId: number) => void;
    setDrivers: (drivers: MarkerData[]) => void;
    clearSelectedDriver: () => void;
}

declare interface DriverCardProps {
    item: MarkerData;
    selected: number;
    setSelected: () => void;
}

declare interface TrialButtonProps{
    textStyle?:string;
    containStyle?:string;
}

interface EmergencyStore {
    emergencyStatus: string | null; // Status can be a string or null
    setEmergencyStatus: (status: string | null) => void; // Function to update the status
  }